requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
flask_sqlalchemy==3.1.0
flask_cors==3.0.10
